Modulo de Ajustes y Cambio en configuraciones de google drive

+++ database/migrations/2022_02_08_174002_create_companies_table.php
@@ -1,6 +1,7 @@
 <?php
 
 use App\Models\User;
+use App\Models\Setting;
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
@@ -38,15 +39,24 @@ class CreateCompaniesTable extends Migration
      */
     public function down()
     {
-        $path='/';
-        $recursive=false;
-        $contents=collect(Storage::disk('google')->listContents($path, $recursive));
+        $setting=Setting::where('id', 1)->first();
+        if (!is_null($setting)) {
+            config(['filesystems.disks.google.clientId' => $setting->google_drive_client_id, 'filesystems.disks.google.clientSecret' => $setting->google_drive_client_secret, 'filesystems.disks.google.refreshToken' => $setting->google_drive_refresh_token, 'filesystems.disks.google.folderId' => $setting->google_drive_folder_id]);
 
-        $customers=User::where('id', '!=', '1')->get();
-        foreach ($customers as $customer) {
-            $directory=$contents->where('type', '=', 'dir')->where('filename', '=', $customer->slug)->first();
-            if ($directory) {
-                Storage::disk('google')->deleteDirectory($directory['path']);
+            try {
+                $path='/';
+                $recursive=false;
+                $contents=collect(Storage::disk('google')->listContents($path, $recursive));
+
+                $customers=User::where('id', '!=', '1')->get();
+                foreach ($customers as $customer) {
+                    $directory=$contents->where('type', '=', 'dir')->where('filename', '=', $customer->slug)->first();
+                    if ($directory) {
+                        Storage::disk('google')->deleteDirectory($directory['path']);
+                    }
+                }
+            } catch (Exception $e) {
+                Log::error("Google API Exception: ".$e->getMessage());
             }
         }
